rootfs: arm: Add more test targets
Test vexpress-a15 with qemu_arm_vexpress_defconfig,
and vexpress-a9/vexpress-a15 with multi_v7_defconfig.

+++ rootfs/scripts/common.sh
@@ -86,7 +86,10 @@ dosetup()
     doclean ${ARCH}
 
     mkdir -p arch/${xARCH}/configs
-    cp ${progdir}/${defconfig} arch/${xARCH}/configs
+    if [ -e ${progdir}/${defconfig} ]
+    then
+        cp ${progdir}/${defconfig} arch/${xARCH}/configs
+    fi
 
     make ARCH=${ARCH} CROSS_COMPILE=${PREFIX} ${defconfig} >/dev/null 2>&1
     if [ $? -ne 0 ]
